Moved the autosuggest helpers to their own file
