+++ src/aims_ui/static/js/e_all_pages/save_and_restore_input_helpers/save_input_listeners.mjs
@@ -1,7 +1,9 @@
 import { getPagePreviouslySearchedValues,
  setPreviouslyStoredValuesForThisPage } from './set_and_get_data.mjs'
 
-function saveValueOfInput(inputId, inputValue, page_name) {
+import { addEventListenerToTriggerSaveOnChangeForAutosuggestComponent } from './autosuggest_listeners.mjs';
+
+export function saveValueOfInput(inputId, inputValue, page_name) {
   console.log(`Saving value of input ${inputId}: ${inputValue}`);
 
   // Firstly get the current pages previously stored input values
@@ -18,71 +20,6 @@ function saveValueOfInput(inputId, inputValue, page_name) {
   setPreviouslyStoredValuesForThisPage(mergedValues, page_name);
 }
 
-// Remove markup from inside an element (i.e for the autosuggest suggestions)
-function removeMarkupFromInsideElement(element) {
-  if (!element) return '';
-
-  const clone = element.cloneNode(true);
-
-  // Replace <strong> with just its text content
-  clone.querySelectorAll('strong').forEach(strongEl => {
-    const textNode = document.createTextNode(strongEl.textContent);
-    strongEl.replaceWith(textNode);
-  });
-
-  // Remove all <span>s entirely
-  clone.querySelectorAll('span').forEach(spanEl => {
-    spanEl.remove();
-  });
-
-  // Return the pure text
-  return clone.textContent.trim();
-}
-
-function addEventListenerToTriggerSaveOnChangeForAutosuggestComponent(inputElement, page_name) {
-  // Given an input element that's an ONS autosuggest component
-  // the event listener must be on the *Suggestions* not the element
-  // as the element itself triggers events like input, change, blur before the value is changed
-  const completeContainerForClassificationFilter = document.querySelector('#complete-container-for-classificationfilter');
-
-  const classNameForSuggestionContainer = "ons-autosuggest__results";
-  const suggestionContainer = completeContainerForClassificationFilter.querySelector('.' + classNameForSuggestionContainer);
-
-  if (!suggestionContainer) { 
-    console.log('No suggestion container found for autosuggest component');
-    return;
-  }
-
-  suggestionContainer.addEventListener('click', event => {
-    const clickedEl = event.target.closest('li');
-
-    if (clickedEl) {
-      const textFromAutosuggest = removeMarkupFromInsideElement(clickedEl);
-
-      // Now save this value
-      saveValueOfInput(inputElement.id, textFromAutosuggest, page_name);
-      console.log('Saved value from autosuggest:', textFromAutosuggest);
-    } else {
-      console.log('Clicked outside of a suggestion <li>');
-    }
-  });
-
-  // Add an event listener for keyboard selection (Enter key)
-  completeContainerForClassificationFilter.addEventListener('keydown', event => {
-    if (event.key === 'Enter') {
-      // The suggestion that is currently focused will have ons-autosuggest__option--focused
-      const focusedSuggestion = suggestionContainer.querySelector('.ons-autosuggest__option--focused');
-      if (focusedSuggestion) {
-        const textFromAutosuggest = removeMarkupFromInsideElement(focusedSuggestion);
-
-        // Now save this value
-        saveValueOfInput(inputElement.id, textFromAutosuggest, page_name);
-        console.log('Saved value from autosuggest (keyboard):', textFromAutosuggest);
-      }
-    }
-  });
-}
-
 export function addEventListenersToTriggerSaveOnChange(saveAndRestoreInputIds, page_name) {
   // Loop over all the ids
   for (const id of saveAndRestoreInputIds) {
@@ -96,8 +33,11 @@ export function addEventListenersToTriggerSaveOnChange(saveAndRestoreInputIds, p
       // Add an event listener to save from a suggestion - STILL REQUIRES A REGULAR INPUT LISTENER
       addEventListenerToTriggerSaveOnChangeForAutosuggestComponent(inputElement, page_name);
     } 
+
+    // Handle regular text 'input's
     inputElement.addEventListener('input', () => {
       saveValueOfInput(inputElement.id, inputElement.value, page_name);
     });
+
   }
 }
